refactored offer component
